617: finalise - docs, logging rule, and detect-secrets hygiene
- research.md: mark HMAC-hash decision resolved
- AGENTS.md/CLAUDE.md: add logging PII/secrets rule for future code
- move get_secret_key tests to test_log_redaction.py (keep test_config.py
  untouched so its pre-existing test secrets are not re-scanned)
- allowlist pre-existing test/default secret-key literals in config.py
